[cleanup] move auth & nav container to AppNavigator component
